+++ src/mpi/attr/comm_delete_attr.c
@@ -24,57 +24,6 @@ int MPI_Comm_delete_attr(MPI_Comm comm, int comm_keyval)
 #undef MPI_Comm_delete_attr
 #define MPI_Comm_delete_attr PMPI_Comm_delete_attr
 
-int MPIR_Comm_delete_attr_impl(MPIR_Comm * comm_ptr, MPII_Keyval * keyval_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Attribute *p, **old_p;
-
-    /* Look for attribute.  They are ordered by keyval handle */
-
-    old_p = &comm_ptr->attributes;
-    p = comm_ptr->attributes;
-    while (p) {
-        if (p->keyval->handle == keyval_ptr->handle) {
-            break;
-        }
-        old_p = &p->next;
-        p = p->next;
-    }
-
-    /* We can't unlock yet, because we must not free the attribute until
-     * we know whether the delete function has returned with a 0 status
-     * code */
-
-    if (p) {
-        int in_use;
-
-        /* Run the delete function, if any, and then free the
-         * attribute storage.  Note that due to an ambiguity in the
-         * standard, if the usr function returns something other than
-         * MPI_SUCCESS, we should either return the user return code,
-         * or an mpich error code.  The precedent set by the Intel
-         * test suite says we should return the user return code.  So
-         * we must not ERR_POP here. */
-        mpi_errno = MPIR_Call_attr_delete(comm_ptr->handle, p);
-        if (mpi_errno)
-            goto fn_fail;
-
-        /* We found the attribute.  Remove it from the list */
-        *old_p = p->next;
-        /* Decrement the use of the keyval */
-        MPII_Keyval_release_ref(p->keyval, &in_use);
-        if (!in_use) {
-            MPIR_Handle_obj_free(&MPII_Keyval_mem, p->keyval);
-        }
-        MPID_Attr_free(p);
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
