+++ src/mpi/attr/comm_get_attr.c
@@ -24,21 +24,11 @@ int MPI_Comm_get_attr(MPI_Comm comm, int comm_keyval, void *attribute_val, int *
 #undef MPI_Comm_get_attr
 #define MPI_Comm_get_attr PMPI_Comm_get_attr
 
-/* Find the requested attribute.  If it exists, return either the attribute
-   entry or the address of the entry, based on whether the request is for
-   a pointer-valued attribute (C or C++) or an integer-valued attribute
-   (Fortran, either 77 or 90).
-
-   If the attribute has the same type as the request, it is returned as-is.
-   Otherwise, the address of the attribute is returned.
-*/
-int MPII_Comm_get_attr(MPI_Comm comm, int comm_keyval, void *attribute_val,
-                       int *flag, MPIR_Attr_type outAttrType)
+int MPII_Comm_get_attr(MPI_Comm comm, int comm_keyval, void *attribute_val, int *flag,
+                       MPIR_Attr_type attr_type)
 {
     int mpi_errno = MPI_SUCCESS;
     MPIR_Comm *comm_ptr = NULL;
-    static PreDefined_attrs attr_copy;  /* Used to provide a copy of the
-                                         * predefined attributes */
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPII_COMM_GET_ATTR);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
@@ -77,180 +67,16 @@ int MPII_Comm_get_attr(MPI_Comm comm, int comm_keyval, void *attribute_val,
 #endif /* HAVE_ERROR_CHECKING */
 
     /* ... body of routine ...  */
-
-    /* Check for builtin attribute */
-    /* This code is ok for correct programs, but it would be better
-     * to copy the values from the per-process block and pass the user
-     * a pointer to a copy */
-    /* Note that if we are called from Fortran, we must return the values,
-     * not the addresses, of these attributes */
-    if (HANDLE_IS_BUILTIN(comm_keyval)) {
-        int attr_idx = comm_keyval & 0x0000000f;
-        void **attr_val_p = (void **) attribute_val;
-#ifdef HAVE_FORTRAN_BINDING
-        /* This is an address-sized int instead of a Fortran (MPI_Fint)
-         * integer because, even for the Fortran keyvals, the C interface is
-         * used which stores the result in a pointer (hence we need a
-         * pointer-sized int).  Thus we use intptr_t instead of MPI_Fint.
-         * On some 64-bit plaforms, such as Solaris-SPARC, using an MPI_Fint
-         * will cause the value to placed into the high, rather than low,
-         * end of the output value. */
-#endif
-        *flag = 1;
-
-        /* FIXME : We could initialize some of these here; only tag_ub is
-         * used in the error checking. */
-        /*
-         * The C versions of the attributes return the address of a
-         * *COPY* of the value (to prevent the user from changing it)
-         * and the Fortran versions provide the actual value (as an Fint)
-         */
-        attr_copy = MPIR_Process.attrs;
-        switch (attr_idx) {
-            case 1:    /* TAG_UB */
-            case 2:
-                *attr_val_p = &attr_copy.tag_ub;
-                break;
-            case 3:    /* HOST */
-            case 4:
-                *attr_val_p = &attr_copy.host;
-                break;
-            case 5:    /* IO */
-            case 6:
-                *attr_val_p = &attr_copy.io;
-                break;
-            case 7:    /* WTIME */
-            case 8:
-                *attr_val_p = &attr_copy.wtime_is_global;
-                break;
-            case 9:    /* UNIVERSE_SIZE */
-            case 10:
-                /* This is a special case.  If universe is not set, then we
-                 * attempt to get it from the device.  If the device is doesn't
-                 * supply a value, then we set the flag accordingly.  Note that
-                 * we must initialize the cached value, in MPIR_Process.attr,
-                 * not the value in the copy, since that is vulnerable to being
-                 * overwritten by the user, and because we always re-initialize
-                 * the copy from the cache (yes, there was a time when the code
-                 * did the wrong thing here). */
-                if (attr_copy.universe >= 0) {
-                    *attr_val_p = &attr_copy.universe;
-                } else if (attr_copy.universe == MPIR_UNIVERSE_SIZE_NOT_AVAILABLE) {
-                    *flag = 0;
-                } else {
-                    /* We call MPID_Get_universe_size only once (see 10.5.1).
-                     * Thus, we must put the value into the "main" copy */
-                    mpi_errno = MPID_Get_universe_size(&MPIR_Process.attrs.universe);
-                    /* --BEGIN ERROR HANDLING-- */
-                    if (mpi_errno != MPI_SUCCESS) {
-                        MPIR_Process.attrs.universe = MPIR_UNIVERSE_SIZE_NOT_AVAILABLE;
-                        goto fn_fail;
-                    }
-                    /* --END ERROR HANDLING-- */
-                    attr_copy.universe = MPIR_Process.attrs.universe;
-                    if (attr_copy.universe >= 0) {
-                        *attr_val_p = &attr_copy.universe;
-                    } else {
-                        *flag = 0;
-                    }
-                }
-                break;
-            case 11:   /* LASTUSEDCODE */
-            case 12:
-                *attr_val_p = &attr_copy.lastusedcode;
-                break;
-            case 13:   /* APPNUM */
-            case 14:
-                /* This is another special case.  If appnum is negative,
-                 * we take that as indicating no value of APPNUM, and set
-                 * the flag accordingly */
-                if (attr_copy.appnum < 0) {
-                    *flag = 0;
-                } else {
-                    *attr_val_p = &attr_copy.appnum;
-                }
-                break;
-        }
-        /* All of the predefined attributes are stored internally as C ints;
-         * since we've set the output value as the pointer to these, we need
-         * to dereference it here. We must also be very careful of the 3
-         * different output cases, since the two Fortran cases correspond
-         * to MPI_Fint and MPIR_FAint (an internal MPICH typedef for C
-         * version of INTEGER (KIND=MPI_ADDRESS_KIND)) */
-        if (*flag) {
-            /* Use the internal pointer-sized-int for systems (e.g., BG/P)
-             * that define MPI_Aint as a different size than intptr_t.
-             * The casts must be as they are:
-             * On the right, the value is a pointer to an int, so to
-             * get the correct value, we need to extract the int.
-             * On the left, the output type is given by the argument
-             * outAttrType - and the cast must match the intended results */
-            /* FIXME: This code is broken.  The MPIR_ATTR_INT is for Fortran
-             * MPI_Fint types, not int, and MPIR_ATTR_AINT is for Fortran
-             * INTEGER(KIND=MPI_ADDRESS_KIND), which is probably an MPI_Aint,
-             * and intptr_t is for exactly the case where MPI_Aint is not
-             * the same as intptr_t.
-             * This code needs to be fixed in every place that it occurs
-             * (i.e., see the win and type get_attr routines). */
-            if (outAttrType == MPIR_ATTR_AINT)
-                *(intptr_t *) attr_val_p = *(int *) *(void **) attr_val_p;
-            else if (outAttrType == MPIR_ATTR_INT) {
-                /* *(int*)attr_val_p = *(int *)*(void **)attr_val_p; */
-                /* This is correct, because the cooresponding code
-                 * in the Fortran interface expects to find a pointer-length
-                 * integer value.  Thus, this works for both big and little
-                 * endian systems. Any changes made here must have
-                 * corresponding changes in src/binding/f77/attr_getf.c ,
-                 * which is generated by src/binding/f77/buildiface . */
-                *(intptr_t *) attr_val_p = *(int *) *(void **) attr_val_p;
-            }
-        }
-    } else {
-        MPIR_Attribute *p = comm_ptr->attributes;
-
-        /*   */
-        *flag = 0;
-        while (p) {
-            if (p->keyval->handle == comm_keyval) {
-                *flag = 1;
-                if (outAttrType == MPIR_ATTR_PTR) {
-                    if (p->attrType == MPIR_ATTR_INT) {
-                        /* This is the tricky case: if the system is
-                         * bigendian, and we have to return a pointer to
-                         * an int, then we may need to point to the
-                         * correct location in the word. */
-#if defined(WORDS_LITTLEENDIAN) || (SIZEOF_VOID_P == SIZEOF_INT)
-                        *(void **) attribute_val = &(p->value);
-#else
-                        int *p_loc = (int *) &(p->value);
-#if SIZEOF_VOID_P == 2 * SIZEOF_INT
-                        p_loc++;
-#else
-#error Expected sizeof(void*) to be either sizeof(int) or 2*sizeof(int)
-#endif
-                        *(void **) attribute_val = p_loc;
-#endif
-                    } else if (p->attrType == MPIR_ATTR_AINT) {
-                        *(void **) attribute_val = &(p->value);
-                    } else {
-                        *(void **) attribute_val = (void *) (intptr_t) (p->value);
-                    }
-                } else {
-                    *(void **) attribute_val = (void *) (intptr_t) (p->value);
-                }
-
-                break;
-            }
-            p = p->next;
-        }
-    }
+    /* Instead, ask for a desired type. */
+    mpi_errno = MPIR_Comm_get_attr_impl(comm_ptr, comm_keyval, attribute_val, flag, attr_type);
+    if (mpi_errno)
+        goto fn_fail;
     /* ... end of body of routine ... */
 
   fn_exit:
     MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
     MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPII_COMM_GET_ATTR);
     return mpi_errno;
-
   fn_fail:
     /* --BEGIN ERROR HANDLING-- */
 #ifdef HAVE_ERROR_CHECKING
@@ -261,9 +87,9 @@ int MPII_Comm_get_attr(MPI_Comm comm, int comm_keyval, void *attribute_val,
                                  comm_keyval, attribute_val, flag);
     }
 #endif
+    /* --END ERROR HANDLING-- */
     mpi_errno = MPIR_Err_return_comm(comm_ptr, __func__, mpi_errno);
     goto fn_exit;
-    /* --END ERROR HANDLING-- */
 }
 
 #endif /* MPICH_MPI_FROM_PMPI */
