+++ src/mpi/attr/type_get_attr.c
@@ -24,12 +24,11 @@ int MPI_Type_get_attr(MPI_Datatype datatype, int type_keyval, void *attribute_va
 #undef MPI_Type_get_attr
 #define MPI_Type_get_attr PMPI_Type_get_attr
 
-int MPII_Type_get_attr(MPI_Datatype datatype, int type_keyval, void *attribute_val,
-                       int *flag, MPIR_Attr_type outAttrType)
+int MPII_Type_get_attr(MPI_Datatype datatype, int type_keyval, void *attribute_val, int *flag,
+                       MPIR_Attr_type attr_type)
 {
     int mpi_errno = MPI_SUCCESS;
     MPIR_Datatype *type_ptr = NULL;
-    MPIR_Attribute *p;
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPII_TYPE_GET_ATTR);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
@@ -68,64 +67,30 @@ int MPII_Type_get_attr(MPI_Datatype datatype, int type_keyval, void *attribute_v
 #endif /* HAVE_ERROR_CHECKING */
 
     /* ... body of routine ...  */
-
-    *flag = 0;
-    p = type_ptr->attributes;
-    while (p) {
-        if (p->keyval->handle == type_keyval) {
-            *flag = 1;
-            if (outAttrType == MPIR_ATTR_PTR) {
-                if (p->attrType == MPIR_ATTR_INT) {
-                    /* This is the tricky case: if the system is
-                     * bigendian, and we have to return a pointer to
-                     * an int, then we may need to point to the
-                     * correct location in the word. */
-#if defined(WORDS_LITTLEENDIAN) || (SIZEOF_VOID_P == SIZEOF_INT)
-                    *(void **) attribute_val = &(p->value);
-#else
-                    int *p_loc = (int *) &(p->value);
-#if SIZEOF_VOID_P == 2 * SIZEOF_INT
-                    p_loc++;
-#else
-#error Expected sizeof(void*) to be either sizeof(int) or 2*sizeof(int)
-#endif
-                    *(void **) attribute_val = p_loc;
-#endif
-                } else if (p->attrType == MPIR_ATTR_AINT) {
-                    *(void **) attribute_val = &(p->value);
-                } else {
-                    *(void **) attribute_val = (void *) (intptr_t) (p->value);
-                }
-            } else
-                *(void **) attribute_val = (void *) (intptr_t) (p->value);
-
-            break;
-        }
-        p = p->next;
-    }
+    mpi_errno = MPIR_Type_get_attr_impl(type_ptr, type_keyval, attribute_val, flag, attr_type);
+    if (mpi_errno)
+        goto fn_fail;
 
     /* ... end of body of routine ... */
 
-#ifdef HAVE_ERROR_CHECKING
   fn_exit:
-#endif
     MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPII_TYPE_GET_ATTR);
     MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
     return mpi_errno;
 
+  fn_fail:
     /* --BEGIN ERROR HANDLING-- */
 #ifdef HAVE_ERROR_CHECKING
-  fn_fail:
     {
         mpi_errno =
             MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
                                  "**mpi_type_get_attr", "**mpi_type_get_attr %D %d %p %p",
                                  datatype, type_keyval, attribute_val, flag);
     }
-    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
-    goto fn_exit;
 #endif
     /* --END ERROR HANDLING-- */
+    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
+    goto fn_exit;
 }
 #endif
