+++ src/mpi/attr/win_create_keyval.c
@@ -61,7 +61,6 @@ int MPI_Win_create_keyval(MPI_Win_copy_attr_function * win_copy_attr_fn,
                           int *win_keyval, void *extra_state)
 {
     int mpi_errno = MPI_SUCCESS;
-    MPII_Keyval *keyval_ptr;
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_WIN_CREATE_KEYVAL);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
@@ -82,28 +81,11 @@ int MPI_Win_create_keyval(MPI_Win_copy_attr_function * win_copy_attr_fn,
 
     /* ... body of routine ...  */
 
-    keyval_ptr = (MPII_Keyval *) MPIR_Handle_obj_alloc(&MPII_Keyval_mem);
-    MPIR_ERR_CHKANDJUMP1(!keyval_ptr, mpi_errno, MPI_ERR_OTHER, "**nomem",
-                         "**nomem %s", "MPII_Keyval");
-    /* Initialize the attribute dup function */
-    if (!MPIR_Process.attr_dup) {
-        MPIR_Process.attr_dup = MPIR_Attr_dup_list;
-        MPIR_Process.attr_free = MPIR_Attr_delete_list;
+    mpi_errno = MPIR_Win_create_keyval_impl(win_copy_attr_fn, win_delete_attr_fn, win_keyval,
+                                            extra_state);
+    if (mpi_errno) {
+        goto fn_fail;
     }
-
-    /* The handle encodes the keyval kind.  Modify it to have the correct
-     * field */
-    keyval_ptr->handle = (keyval_ptr->handle & ~(0x03c00000)) | (MPIR_WIN << 22);
-    MPIR_Object_set_ref(keyval_ptr, 1);
-    keyval_ptr->was_freed = 0;
-    keyval_ptr->kind = MPIR_WIN;
-    keyval_ptr->extra_state = extra_state;
-    keyval_ptr->copyfn.user_function = win_copy_attr_fn;
-    keyval_ptr->copyfn.proxy = MPII_Attr_copy_c_proxy;
-    keyval_ptr->delfn.user_function = win_delete_attr_fn;
-    keyval_ptr->delfn.proxy = MPII_Attr_delete_c_proxy;
-
-    MPIR_OBJ_PUBLISH_HANDLE(*win_keyval, keyval_ptr->handle);
     /* ... end of body of routine ... */
 
   fn_exit:
