+++ src/mpi/attr/win_set_attr.c
@@ -24,12 +24,11 @@ int MPI_Win_set_attr(MPI_Win win, int win_keyval, void *attribute_val)
 #undef MPI_Win_set_attr
 #define MPI_Win_set_attr PMPI_Win_set_attr
 
-int MPII_Win_set_attr(MPI_Win win, int win_keyval, void *attribute_val, MPIR_Attr_type attrType)
+int MPII_Win_set_attr(MPI_Win win, int win_keyval, void *attribute_val, MPIR_Attr_type attr_type)
 {
     int mpi_errno = MPI_SUCCESS;
     MPIR_Win *win_ptr = NULL;
     MPII_Keyval *keyval_ptr = NULL;
-    MPIR_Attribute *p, **old_p;
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPII_WIN_SET_ATTR);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
@@ -73,66 +72,11 @@ int MPII_Win_set_attr(MPI_Win win, int win_keyval, void *attribute_val, MPIR_Att
     }
 #endif /* HAVE_ERROR_CHECKING */
 
+
     /* ... body of routine ...  */
-
-    /* Look for attribute.  They are ordered by keyval handle.  This uses
-     * a simple linear list algorithm because few applications use more than a
-     * handful of attributes */
-
-    old_p = &win_ptr->attributes;
-    p = win_ptr->attributes;
-    while (p) {
-        if (p->keyval->handle == keyval_ptr->handle) {
-            /* If found, call the delete function before replacing the
-             * attribute */
-            mpi_errno = MPIR_Call_attr_delete(win, p);
-            /* --BEGIN ERROR HANDLING-- */
-            if (mpi_errno) {
-                /* FIXME : communicator of window? */
-                goto fn_fail;
-            }
-            /* --END ERROR HANDLING-- */
-            p->value = (MPII_Attr_val_t) (intptr_t) attribute_val;
-            p->attrType = attrType;
-            /* Does not change the reference count on the keyval */
-            break;
-        } else if (p->keyval->handle > keyval_ptr->handle) {
-            MPIR_Attribute *new_p = MPID_Attr_alloc();
-            MPIR_ERR_CHKANDJUMP1(!new_p, mpi_errno, MPI_ERR_OTHER,
-                                 "**nomem", "**nomem %s", "MPIR_Attribute");
-            new_p->keyval = keyval_ptr;
-            new_p->attrType = attrType;
-            new_p->pre_sentinal = 0;
-            new_p->value = (MPII_Attr_val_t) (intptr_t) attribute_val;
-            new_p->post_sentinal = 0;
-            new_p->next = p->next;
-            MPII_Keyval_add_ref(keyval_ptr);
-            p->next = new_p;
-            break;
-        }
-        old_p = &p->next;
-        p = p->next;
-    }
-    if (!p) {
-        MPIR_Attribute *new_p = MPID_Attr_alloc();
-        MPIR_ERR_CHKANDJUMP1(!new_p, mpi_errno, MPI_ERR_OTHER,
-                             "**nomem", "**nomem %s", "MPIR_Attribute");
-        /* Did not find in list.  Add at end */
-        new_p->attrType = attrType;
-        new_p->keyval = keyval_ptr;
-        new_p->pre_sentinal = 0;
-        new_p->value = (MPII_Attr_val_t) (intptr_t) attribute_val;
-        new_p->post_sentinal = 0;
-        new_p->next = 0;
-        MPII_Keyval_add_ref(keyval_ptr);
-        *old_p = new_p;
-    }
-
-    /* Here is where we could add a hook for the device to detect attribute
-     * value changes, using something like
-     * MPID_Win_attr_hook(win_ptr, keyval, attribute_val);
-     */
-
+    mpi_errno = MPIR_Win_set_attr_impl(win_ptr, keyval_ptr, attribute_val, attr_type);
+    if (mpi_errno)
+        goto fn_fail;
     /* ... end of body of routine ... */
 
   fn_exit:
@@ -150,9 +94,9 @@ int MPII_Win_set_attr(MPI_Win win, int win_keyval, void *attribute_val, MPIR_Att
                                  win_keyval, attribute_val);
     }
 #endif
+    /* --END ERROR HANDLING-- */
     mpi_errno = MPIR_Err_return_win(win_ptr, __func__, mpi_errno);
     goto fn_exit;
-    /* --END ERROR HANDLING-- */
 }
 #endif
